+++ hephaistos-data/lua/FixLuaState/UIData.lua
@@ -15,6 +15,3 @@ TraitUI.IconStartY = Hephaistos.RecomputeFixedYFromCenter(TraitUI.IconStartY)
 
 ConsumableUI.StartX = Hephaistos.RecomputeFixedXFromRight(ConsumableUI.StartX)
 ConsumableUI.StartY =  Hephaistos.RecomputeFixedYFromBottom(ConsumableUI.StartY)
-
-GunUI.StartX = Hephaistos.RecomputeFixedXFromCenter(GunUI.StartX)
-GunUI.StartY = Hephaistos.RecomputeFixedYFromCenter(GunUI.StartY)