+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix locked keepsakes icons.
+
 ## [v0.8.0] - 2021-08-19
 
 ### Fixed