+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fix fire animation when getting by lava.
 -   Fix top left icon on boon choice menu.
 -   Fix reroll vignette overlay.
+-   Fix gun ammo UI begin offset to the right.
 
 ## [v0.8.0] - 2021-08-19