+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.9.0] - 2021-08-21
+
 ### Fixed
 
 -   Fix locked keepsakes icons.
@@ -97,7 +99,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v0.9.0...HEAD
+
+[v0.9.0]: https://github.com/nbusseneau/hephaistos/compare/v0.8.0...v0.9.0
 
 [v0.8.0]: https://github.com/nbusseneau/hephaistos/compare/v0.7.0...v0.8.0